Changes after PR feedback: - Adding back Configuration file/class - Making BatchSignInput a wrapper around Immutable properties - Fixes around coding convention
